requirements.txt

python-dotenv==1.1.0
google-api-python-client==2.172.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
logging==0.4.9.6
datetime==5.5